refactor(server): extract search boot to src/core/search-startup.ts
First safe slice of the server.ts split. The full split is high-risk (module-graph
contamination — registerCorePlugins/setEmbeddedAssets must stay in the entry; strict
boot-ordering contracts; exact route-precedence reproduction), so the self-contained,
unit-testable search-boot logic comes out first and the dangerous router/boot-ordering
seams are left for a dedicated effort.

Moves runSearchStartupBootstrap + the SEARCH_STARTUP_RETRY_MS deferred-retry wiring out of
server.ts into src/core/search-startup.ts, beside search-migration.ts / search-registry.ts.
server.ts now calls a single bootSearch(migration); the dynamic import of search-registry
stays dynamic (graph-pruning). Logger allowlist gains 'search-startup'. server.ts: 833 → 777.

Verified: typecheck + lint clean; full suite 5105/0; full 3-target binary build; and an
isolated-home boot smoke of the built binary — the server boots and the search tables
register through the extracted bootSearch path (the only log errors are environmental: an
empty OPENCLAW_HOME means the schedule plugin's cron activation/reindex can't run, and
Antfly's transient shard-scan on freshly-created empty tables — neither from this change).
